+++ tests/testthat/test-vroom.R
@@ -598,3 +598,11 @@ test_that("escaped/quoted comments are ignored", {
   expect_equal(out1$x, "#")
   expect_equal(out2$x, "#")
 })
+
+test_that("name repair with custom functions works", {
+  add_y <- function(x) {
+    paste(x, "y", sep = "_")
+  }
+  out <- vroom(I("x,y,z\n1,2,3"), .name_repair = add_y)
+  expect_equal(colnames(out), c("x_y", "y_y", "z_y"))
+})